[refactor] 수동과 자동 분리
